PRR-126 [refactor] modify nav constants structure
