+++ src/constants/navigation/index.ts
@@ -1,0 +1,2 @@
+export * from './global';
+export * from './mypage';
